[TNT-151] move: 역할 선택 화면 roleselect 모듈로 이동
- 이미지도 모듈 내로 이동
