+++ core/designsystem/src/main/res/values/strings.xml
@@ -4,13 +4,4 @@
     <string name="text_counter">%1$d/%2$d자</string>
     <string name="placeholder_content_input">내용을 입력해주세요</string>
 
-    <!--  Sign Up  -->
-    <string name="next">다음</string>
-    <string name="start">시작하기</string>
-    <string name="trainer">트레이너</string>
-    <string name="trainee">트레이니</string>
-
-    <string name="signup_select_role_title">안녕하세요!\n어떤 회원으로 이용하시겠어요?</string>
-    <string name="signup_select_role_subtitle">트레이너와 트레이니 중 선택해주세요.</string>
-
 </resources>
