+++ feature/roleselect/src/main/res/drawable/img_select_role_trainer.xml
